re-org of code to improve perf
