+++ src/components/translatable/Translatable.js
@@ -8,15 +8,25 @@ import RowHeader from './RowHeader';
 import { FilesHeader } from '../files-header';
 import { AppContext } from '../../App.context';
 import { TargetFileContext } from '../../core/TargetFile.context';
+import { ResourcesContextProvider } from "scripture-resources-rcl";
+
+import { testament } from '../../core/bcv.js';
+import { SERVER_URL } from '../../core/state.defaults';
+
 
 function Translatable() {
   const classes = useStyles();
   const [wrapperElement, setWrapperElement] = useState(null);
+
+  // manage the state of the resources for the provider context
+  const [ resources, setResources ] = React.useState([]);
+
   const {
     state: {
       language, sourceRepository, targetRepository, filepath,
     },
   } = useContext(AppContext);
+
   const { state: sourceFile } = useContext(FileContext);
 
   const { state: targetFile, actions: targetFileActions } = useContext(TargetFileContext);
@@ -26,6 +36,7 @@ function Translatable() {
       window.scrollTo(0, wrapperElement.offsetParent.offsetTop);
     }
   }, [wrapperElement]);
+
 
   const translatableComponent = useMemo(() => {
     let _translatable = <h3>Unsupported File. Please select .md or .tsv files.</h3>;
@@ -38,6 +49,7 @@ function Translatable() {
         };
         _translatable = <MarkDownTranslatable {...translatableProps} />;
       } else if (sourceFile.filepath.match(/\.tsv$/)) {
+        const bookId = sourceFile.filepath.split(/\d+-/)[1];
         const delimiters = { row: '\n', cell: '\t' };
         const rowHeader = (rowData, actionsMenu) => (
           <RowHeader rowData={rowData} actionsMenu={actionsMenu} delimiters={delimiters} />
@@ -55,11 +67,43 @@ function Translatable() {
           delimiters,
           config,
         };
-        _translatable = <DataTable {...translatableProps} />;
+        const reference = { bookId };
+        const _testament = testament(reference);
+        let hebrewLink = 'unfoldingWord/hbo/uhb/master';
+        let greekLink = 'unfoldingWord/el-x-koine/ugnt/master';
+        let originalLink = (_testament === 'old') ? hebrewLink : greekLink;
+      
+        // need to add reference bookId to resource links
+        const _resourceLinks = [
+          originalLink,
+          'unfoldingWord/en/ult/master',
+          'unfoldingWord/en/ust/master',
+        ];
+        const resourceLinks = _resourceLinks.map( (link) => {
+          return link+'/'+bookId;
+        });
+      
+        const serverConfig = { 
+          server: SERVER_URL,
+          cache: {
+            maxAge: 1 * 1 * 1 * 60 * 1000, // override cache to 1 minute
+          },
+        };
+
+        _translatable = (
+          <ResourcesContextProvider
+            resourceLinks={resourceLinks}
+            resources={resources}
+            onResources={setResources}
+            config={serverConfig}
+          >
+            <DataTable {...translatableProps} />
+          </ResourcesContextProvider>
+        );
       }
     }
     return _translatable;
-  }, [filepath, sourceFile, targetFile, targetFileActions.save]);
+  }, [filepath, sourceFile, targetFile, targetFileActions.save, resources]);
 
   useEffect(() => {
       scrollToTop();
